+++ packages/blocks/tsup.config.ts
@@ -1,14 +1,16 @@
 import { defineConfig } from "tsup";
 
-export default defineConfig([
-  {
-    entry: ["src/index.ts"],
-    format: ["esm", "cjs"],
-    dts: true,
-    sourcemap: true,
-    clean: true,
-    outDir: "dist",
-    minify: true,
-    target: "es2020",
+export default defineConfig({
+  entry: {
+    index: "src/index.ts",
+    shadcn: "src/shadcn.ts",
+    magicui: "src/magicui.ts",
   },
-]);
+  format: ["esm", "cjs"],
+  dts: true,
+  sourcemap: true,
+  clean: true,
+  outDir: "dist",
+  minify: true,
+  target: "es2020",
+});